+++ package.json
@@ -16,7 +16,11 @@
     "react": "16.9.0",
     "react-native": "0.59.9",
     "react-native-firebase": "^5.6.0",
-    "react-viro": "2.17.0"
+    "react-redux": "^7.2.0",
+    "react-redux-firebase": "^3.3.0",
+    "react-viro": "2.17.0",
+    "redux": "^4.0.5",
+    "redux-firestore": "^0.13.0"
   },
   "devDependencies": {
     "@babel/core": "7.8.7",
