+++ store/reducer.js
@@ -1,0 +1,11 @@
+// import { combineReducers } from 'redux'
+// import { firebaseReducer } from 'react-redux-firebase'
+// import { firestoreReducer } from 'redux-firestore'
+
+// export const initialState = {}
+
+
+// export const rootReducer = combineReducers({
+//   firestore: firestoreReducer
+// })
+
